Poids = 1/n

+++ choquet/src/model/Node.java
@@ -4,7 +4,7 @@ import java.util.Vector;
 
 public class Node {
 	private String name;
-	private int weight;
+	private float weight;
 	private Vector<Node> parent;
 	private Vector<Node> child;
 
@@ -14,10 +14,10 @@ public class Node {
 	public void setName(String name) {
 		this.name = name;
 	}
-	public int getWeight() {
+	public float getWeight() {
 		return weight;
 	}
-	public void setWeight(int weight) {
+	public void setWeight(float weight) {
 		this.weight = weight;
 	}
 	public Vector<Node> getParent() {
